ucdXML.py -- first commit
This module provides access to the Unicode Character Database (UCD) that defines character properties for all Unicode characters.
The methods provided here are similar to those provided by python's built-in unicodedata module.   Rather than using a pre-computed
fixed version of Unicode, however, this module requires the user to supply the Unicode Data file as a parameter.